refactor(Transactions): move component to its own directory
